+++ packages/mdx/src/getMDXBySlug.tsx
@@ -1,16 +1,8 @@
 import fs from 'node:fs'
 import path from 'node:path'
 import compareVersions from 'compare-versions'
-import { bundleMDX } from 'mdx-bundler'
-import readingTime from 'reading-time'
-import type { Frontmatter } from './types'
-import { rehypeHighlightCode } from './rehypeHighlightCode'
-import rehypeMetaAttribute from './rehypeMetaAttribute'
-import rehypeAutolinkHeadings from 'rehype-autolink-headings'
-import rehypeSlug from 'rehype-slug'
-import { getHeadings } from './getHeadings'
-
-export type UnifiedPlugin = import('unified').Plugin[]
+import type { Frontmatter, UnifiedPlugin } from './types'
+import { getMDX } from './getMDX'
 
 export const getMDXBySlug = async (
   basePath: string,
@@ -27,33 +19,8 @@ export const getMDXBySlug = async (
 
   const filePath = path.join(basePath, `${mdxPath}.mdx`)
   const source = fs.readFileSync(filePath, 'utf8')
-  const { frontmatter, code } = await getMDX(source, extraPlugins)
-  return {
-    frontmatter: {
-      ...frontmatter,
-      headings: getHeadings(source),
-      readingTime: readingTime(code),
-    } as Frontmatter,
-    code,
-  }
-}
-
-export async function getMDX(source: string, extraPlugins?: UnifiedPlugin) {
-  return await bundleMDX({
-    source,
-    mdxOptions(options) {
-      const plugins = [
-        ...(extraPlugins || []),
-        ...(options.rehypePlugins ?? []),
-        rehypeMetaAttribute,
-        rehypeHighlightCode,
-        rehypeAutolinkHeadings,
-        rehypeSlug,
-      ]
-      options.rehypePlugins = plugins as any
-      return options
-    },
-  })
+  const bundle = await getMDX(source, extraPlugins)
+  return bundle
 }
 
 export function getAllVersionsFromPath(fromPath: string): string[] {
