+++ packages/mdx/types/getMDX.d.ts
@@ -1,0 +1,6 @@
+import type { Frontmatter, UnifiedPlugin } from './types';
+export declare function getMDX(source: string, extraPlugins?: UnifiedPlugin): Promise<{
+    frontmatter: Frontmatter;
+    code: string;
+}>;
+//# sourceMappingURL=getMDX.d.ts.map
