+++ packages/mdx/types/getMDXBySlug.d.ts
@@ -1,20 +1,7 @@
-import type { Frontmatter } from './types';
-export type UnifiedPlugin = import('unified').Plugin[];
+import type { Frontmatter, UnifiedPlugin } from './types';
 export declare const getMDXBySlug: (basePath: string, slug: string, extraPlugins?: UnifiedPlugin) => Promise<{
     frontmatter: Frontmatter;
     code: string;
-}>;
-export declare function getMDX(source: string, extraPlugins?: UnifiedPlugin): Promise<{
-    code: string;
-    frontmatter: {
-        [key: string]: any;
-    };
-    errors: import("esbuild").Message[];
-    matter: Omit<import("gray-matter").GrayMatterFile<string>, "data"> & {
-        data: {
-            [key: string]: any;
-        };
-    };
 }>;
 export declare function getAllVersionsFromPath(fromPath: string): string[];
 //# sourceMappingURL=getMDXBySlug.d.ts.map
