+++ packages/mdx/types/index.d.ts
@@ -1,4 +1,5 @@
 export { getAllFrontmatter } from './getAllFrontmatter';
+export { getMDX } from './getMDX';
 export { getAllVersionsFromPath, getMDXBySlug } from './getMDXBySlug';
 export { createCodeHighlighter } from './highlightCode';
 export type { Frontmatter } from './types';
