+++ packages/mdx/types/types.d.ts
@@ -27,4 +27,5 @@ export type Frontmatter = {
     package?: string;
     demoName?: string;
 };
+export type UnifiedPlugin = import('unified').Plugin[];
 //# sourceMappingURL=types.d.ts.map
